change Int to Double for the timer values and fix the text fields constraints.

+++ src/xcode/ENA/ENA/Source/Models/Exposure/WarnOthersNotificationsTimer.swift
@@ -20,7 +20,7 @@
 import Foundation
 
 /// Defines two timer values for the first and the second notification to be shown. The values are meant to be seconds.
-enum WarnOthersNotificationsTimer: Int {
+enum WarnOthersNotificationsTimer: Double {
 	
 	case timerOneTime = 7200
